feat(entity)將order兩個table分開，並且新增單獨放order_detail table的file
